Allow for translation keys to not be namespaced
I want to be able to define a minimum translation file structure, which means that I should be able to have `fields` and `buttons` properties at the top level. So default to a shared translation namespace of '' instead of 'shared'.

Add some tests for this as well. Tests are obviously short of full coverage, but they're a start.