Inputs need to be linked to form hints (where they exist) by use of `aria-describedby`.
For single input fields, ie passport number, add `aria-describedby`, which will be shown if a hint exists. Add an ID to the hint so that input is linked to this.

For date groups, a hint is shown at a group level and not individually. Update the template mixin to send `hintId`, which is used in another `aria-describedby`. The hint ID will need to be manually added to the group hint for date inputs.