Add error messages to form groups.
Depending on the input type there may be two versions of the error message, this is for visually impaired and sighted users. It is intended that the error message is presented in the correct order depending on the context. Each user will only see or hear the error message. This is achieved by using `aria-hidden="true"` and `.visuallyhidden`. Where there is only one instance of the error message, `aria-describedby` is used. Invalid fields are also marked up with `aria-invalid="true"`.