Fix url path for apps not mounted on a sub-path
Currently the function for res.locals.url assumes req.baseUrl is set. For apps that are not mounted on a sub-path this may not be the case, which can result in the full location of the path on the server being returned for relative urls. This change adds a check to see if req.baseUrl is set.