Use req.translate for translations
This is functionally identical to the previous implementation except that it removes the (currently unused outside its own tests) translate option, and `req.translate` is called at runtime instead of being aliased onto a separate function.

This means that an app only needs to load mixins once, even if the translation function might change in different parts of the app.

In particular, in bootstrap apps it is no longer necessary to mount template mixins separately in all child apps, as well as globally in order to be able to read the relevant translations.